Delay selection when constraint exists
When a node makes it into the graph, it potentially can transform
pending constraints into valid edges. When that happens, we defer the
selection of the the node so that it happens when the constraint
selector also is known.

In the best case, a fixed version is computed, skipping a dynamic
version resolution. In the worst case, the defer was not needed because
the constraint does not change the selection result.